featured users

+++ internal/app/util/rbac.go
@@ -2,10 +2,10 @@ package util
 
 import "fmt"
 
-func FormatRbacSubject(adminID uint64) string {
+func FormatRbacSubject(adminID uint) string {
 	return fmt.Sprintf("user%d", adminID)
 }
 
-func FormatRbacRole(roleID uint64) string {
+func FormatRbacRole(roleID uint) string {
 	return fmt.Sprintf("role%d", roleID)
 }
